requirements.txt
pycaw>=20240210
pywin32>=306
pystray>=0.19.5
Pillow>=10.0.0
schedule>=1.2.0
pyinstaller>=6.0.0
